+++ src/store/modules/survey.js
@@ -1,0 +1,11 @@
+export const survey = {
+  namespaced: true,
+  state: {
+    hintVisible: true
+  },
+  mutations: {
+    setHintVisible (state, visible) {
+      state.hintVisible = visible
+    }
+  }
+}